requirements.txt
python-telegram-bot[job-queue]>=21.0
pydantic-settings>=2.0
pytz>=2024.1
